Move all extra display items to own module
