+++ src/api/rest/comment.rs
@@ -1,7 +1,6 @@
 use serde::{Deserialize, Serialize};
 
 use crate::api::serialize::todoist_rfc3339;
-use owo_colors::OwoColorize;
 
 use super::{ProjectID, TaskID};
 
@@ -57,26 +56,4 @@ pub struct CreateComment {
     /// The text of the comment. Supports markdown.
     pub content: String,
     // TODO: pub attachment: Option<Attachment>,
-}
-
-/// FullComment allows to display full comment metadata when [std::fmt::Display]ing it.
-pub struct FullComment<'a>(pub &'a Comment);
-
-impl std::fmt::Display for FullComment<'_> {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        let FullComment(comment) = self;
-        writeln!(f, "ID: {}", comment.id.bright_yellow())?;
-        writeln!(f, "Posted: {}", comment.posted)?;
-        writeln!(
-            f,
-            "Attachment: {}",
-            if comment.attachment.is_some() {
-                "Yes"
-            } else {
-                "No"
-            }
-        )?;
-        write!(f, "Content: {}", comment.content)?;
-        Ok(())
-    }
 }
